[SERVER] connect db, setup authen, author

+++ Sources/Server/RoomifyAR/RoomifyAR/Entities/Order.cs
@@ -3,7 +3,7 @@
     public class Order : BaseEntityAudit
     {
         public int UserId { get; set; }
-        public decimal TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public string Status { get; set; } = "";
     }
 }
